CC_Project_init/train_management/requirements.txt
Flask-Cors==4.0.0
flask
requests
mysql-connector-python